refactor: rename TreeView --> SimpleTree
