+++ src/DiscoPoP/DiscoPoPSuggestionTree.ts
@@ -3,7 +3,7 @@ import { Suggestion } from './classes/Suggestion/Suggestion'
 import { FileMapping } from '../FileMapping/FileMapping'
 import { DiscoPoPResults } from './classes/DiscoPoPResults'
 import { Commands } from '../Utils/Commands'
-import { SimpleTree, SimpleTreeNode } from '../Utils/TreeViews'
+import { SimpleTree, SimpleTreeNode } from '../Utils/SimpleTree'
 
 /**
  * A suggestion group is a group of suggestions of the same type.
