+++ src/Utils/SimpleTree.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode'
 
 // The following interface and class simplify the creation of TreeViews
 
-export interface SimpleTreeNode<T extends SimpleTreeNode<T>> {
+export interface SimpleTreeNode<T extends SimpleTreeNode<T> | undefined> {
     /** Returns the view for this tree node */
     getView(): vscode.TreeItem
